Add BEM validation tests, calibration generation script, and benchmark data
- Implemented comprehensive tests for the pyhorn_physics.bem_validation module, covering BEM reference data loading, TMM prediction creation, SPL and impedance comparisons, and geometry generation.
- Created a script (bem_generate_calibration.py) to generate BEM radiation calibration JSON files from BEM solver output.
- Added synthetic benchmark data for an exponential horn, including reference SPL data and calibration files.
- Introduced fixture YAML files for driver and horn configurations to facilitate BEM vs TMM validation.

+++ pyhorn_cli/cli/design_wizard.py
@@ -349,15 +349,15 @@ def hornresp(
     data: dict = {
         "enclosure_type": enclosure_type.upper(),
         "sections": [section],
-        "n_segments": int(n_segments),
+        "n_segments": n_segments,
     }
     if lrc > 0 or vrc is not None:
         data["rear_chamber"] = {
-            "lrc": float(lrc),
-            "vrc": float(vrc if vrc is not None else lrc * throat_area),
+            "lrc": lrc,
+            "vrc": vrc if vrc is not None else lrc * throat_area,
         }
     if vtc > 0:
-        data["throat_chamber"] = {"vtc": float(vtc)}
+        data["throat_chamber"] = {"vtc": vtc}
 
     yaml_text = yaml.safe_dump(data, sort_keys=False, default_flow_style=False)
     if output_yaml is not None:
